CLN: deduplicate __setitem__ and _reduce on masked arrays
